% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phd_data.R
\name{phd_data}
\alias{phd_data}
\title{Simulated data based on PhD thesis of P.J.A. Stam (2007)}
\usage{
phd_data(nobs = 100, seed = NULL)
}
\arguments{
\item{nobs}{Number of observations}

\item{seed}{Optional seed number}
}
\value{
A data frame with 63 variables
}
\description{
Simulated data based on PhD thesis of P.J.A. Stam (2007)
}
\examples{
df <- phd_data()            # df w/ 100 observations and random seed
df <- phd_data(1000)        # df w/ 1000 observations and random seed
df <- phd_data(1000, 1234)  # df w/ 1000 observations and seed 1234
}
