% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvedata-package.R
\docType{package}
\name{rvedata-package}
\alias{rvedata}
\alias{rvedata-package}
\title{rvedata: Data Sets for Risk Equalization}
\description{
Datasets for learning and advancing risk equalization, enabling smart decisions and research.
}
\author{
\strong{Maintainer}: Piet Stam \email{pjastam@gmail.com} (\href{https://orcid.org/0000-0003-3680-6391}{ORCID})

}
\keyword{internal}
